// src/types/StyleTypes.ts
import { ViewStyle, TextStyle } from 'react-native';

export interface IHomeStyleTypes {
  backgroundStyle: ViewStyle;
  containers: ViewStyle;
  minMaxContainer: ViewStyle;
  topNav: ViewStyle;
  iconText: TextStyle;
  temperatureText: TextStyle;
  subtitleText: TextStyle;
  topNavContents: TextStyle;
}